+++ R/LocalModel.R
@@ -85,7 +85,6 @@
 #' 
 #' The package `lime` with the original implementation
 #' @export
-#' @importFrom glmnet glmnet
 #' @importFrom stats dist
 #' @examples 
 #' if (require("randomForest")) {
@@ -213,7 +212,7 @@ LocalModel = R6::R6Class("LocalModel",
         geom_col(aes(y = effect, x = reorder(feature.value,effect))) + 
         coord_flip() + 
         ylab("effect") + 
-        xlab("feature.value") 
+        xlab("") 
       if (!private$multiClass) {
         original_prediction = self$predictor$predict(self$x.interest)[[1]]
         p = p + ggtitle(sprintf("Actual prediction: %.2f\nLocalModel prediction: %.2f",