+++ tests/testthat/test-FeatureImp.R
@@ -119,19 +119,47 @@ test_that("Model receives data.frame without additional columns", {
   expect_r6(imp)
 })
 
-
 set.seed(12)
-X = data.frame(x1 = 1:10, x2 = 1:10, x3 = 1:10)
+X = data.frame(x1 = 1:100)
+X$x2 = X$x1 + rnorm(100)
+X$x3 = rnorm(100)
+X$x4 = sample(c(0,1), size = 100, replace = TRUE)
+X$x5 = factor(sample(c(1,2,3), size = 100, replace = TRUE))
 y = X[,1] + X[,2] + rnorm(10, 0, 0.1)
-
-
 pred.fun = function(newdata){
-	newdata[,1] + newdata[,2]
+  newdata[,1] + newdata[,2]
 }
-
 pred = Predictor$new(data = X, predict.fun = pred.fun, y = y)
 
 test_that("Feature Importance 0", {
  fimp = FeatureImp$new(pred, loss = "mae", n.repetitions = 3)
  expect_equal(fimp$results$importance[3], 1) 
 })
+
+
+test_that("Feature Importance 0", {
+ fimp = FeatureImp$new(pred, loss = "mae", n.repetitions = 3)
+ expect_equal(fimp$results$importance[3], 1)
+})
+
+test_that("Conditional Feature Importance", {
+  predictor1 = Predictor$new(data = X, y = y, predict.fun = pred.fun, conditional = TRUE)
+  var.imp = FeatureImp$new(predictor1, loss = "mse", conditional = TRUE)
+  dat = var.imp$results
+  expect_class(dat, "data.frame")
+  expect_false("data.table" %in% class(dat))
+  expect_equal(colnames(dat), expected_colnames)
+  expect_equal(nrow(dat), ncol(X))
+  p = plot(var.imp)
+  expect_s3_class(p, c("gg", "ggplot"))
+  p
+  p = plot(var.imp, sort = FALSE)
+  expect_s3_class(p, c("gg", "ggplot"))
+  p
+  
+  p = var.imp$plot()
+  expect_s3_class(p, c("gg", "ggplot"))
+  p
+})
+
+